Split Switch into two Groups
