+++ components/logicalComponents/BitSwitch.js
@@ -3,7 +3,7 @@ import withRedux              from 'next-redux-wrapper'
 import { connect }            from "react-redux"
 import { initStore }          from '../../store'
 
-import { changeBit, bitgroupAction } from "../../actions/logicalActions";
+import { changeBit } from "../../actions/logicalActions";
 import Button                 from '../logicalComponents/Button';
 import Center                 from '../Center';
 import Slider                 from '../Btn'
@@ -23,34 +23,19 @@ class BitSwitch extends React.Component {
       style={{width: '19px', margin: '0 19.5px'}}/>
     });
 
-    let operations =
-      ["NOT", "SHL", "SHR"].map(function(x) {
-      return index => props.bitgroupAction(name, x)
-    });
-
-    let groupActions = ["!", "<<< shift", "shift >>>"]
-      .map(function(x, index) {
-      return <Button onClick={operations[index]} key={x} text={x.toLowerCase()}/>
-    });
-
     return(
       <div>
-
-      <style jsx>{`
-        .bitButtons {
-          margin: 1em 0px;
-          display: flex;
-          flex-direction: row;
-        }
-      `}</style>
+        <style jsx>{`
+          .bitButtons {
+            margin: 1em 0px;
+            display: flex;
+            flex-direction: row;
+          }
+        `}</style>
 
         <div className="bitButtons">
           {buttons}
         </div>
-
-        <Center className="groupActionButtons">
-          {groupActions}
-        </Center>
       </div>
     )
   }
@@ -61,7 +46,6 @@ export default withRedux(initStore)(
     s => s,
     dispatch => ({
       changeBit: bindActionCreators(changeBit, dispatch),
-      bitgroupAction: bindActionCreators(bitgroupAction, dispatch)
     })
   )(BitSwitch)
 )
